+++ src/navigation/StartStack.js
@@ -5,6 +5,7 @@ import Tutorial from '../screens/Tutorial'
 import EditCronograma from '../screens/EditCronograma'
 import WorkView from '../screens/WorkView'
 import DetailCronograma from '../screens/DetailCronograma'
+import AddCronograma from '../screens/AddCronograma'
 
 const Stack = createStackNavigator()
 
@@ -17,6 +18,7 @@ function StartStack() {
       <Stack.Screen name="Editar" component={EditCronograma} />
       <Stack.Screen name="Trabajar" component={WorkView} />
       <Stack.Screen name="Informacion" component={DetailCronograma} />
+      <Stack.Screen name="AddCronograma" component={AddCronograma} />
     </Stack.Navigator>
   )
 }