+++ src/strings.js
@@ -1,0 +1,6 @@
+export const SCHEDULES_NAME = 'Cronogramas'
+export const SCHEDULES_LIST = 'Listado de Cronogramas'
+export const SCHEDULE_CREATE = 'Crear Cronograma'
+export const CARDS_NAME = 'Tarjetas'
+export const CARDS_LIST = 'Listado de Tarjetas'
+export const CARD_CREATE = 'Crear Tarjeta'